fix(vision): correct match_boxes_to_anchor_grid's classes type annotation
Found via a real mypy run (pyproject.toml has [tool.mypy] configured but
nothing had ever actually run it): the classes param was typed
list[int] | None, but both real call sites pass list[float] (0.0/1.0), and
the function stores it into a torch.float32 tensor consumed by
F.binary_cross_entropy, which needs float targets. Runtime behavior was
already correct -- this fixes the type declaration to match, introduced by
this session's earlier print_mark classification fix (7b5c2dc).

+++ packages/cs_vision/train_rfdetr.py
@@ -64,7 +64,7 @@ def match_boxes_to_anchor_grid(
     anchors_np: np.ndarray,
     boxes: list[list[float]],
     masks: list[np.ndarray],
-    classes: list[int] | None = None,
+    classes: list[float] | None = None,
     num_queries: int = NUM_QUERIES,
     mask_size: int = MASK_SIZE,
 ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor, torch.Tensor]:
